+++ examples/example_pro/src/FirestoreApp/collection_actions/SampleCollectionActions.tsx
@@ -51,7 +51,6 @@ export function CustomFiltersActions({
                     <IconButton size={"small"} onClick={() => updateFilter(null)}>
                         <ClearIcon size={"smallest"}/>
                     </IconButton> : undefined}
-                renderValue={(value) => value}
                 onValueChange={updateFilter}
                 size={"medium"}
                 value={categoryFilterValue}>