+++ packages/cli/vite.config.ts
@@ -2,7 +2,6 @@
 import path from "path";
 
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react"
 
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
@@ -42,6 +41,5 @@ export default defineConfig(() => ({
         }
     },
     plugins: [
-        react({})
     ]
 }));