+++ packages/collection_editor/vite.config.ts
@@ -4,6 +4,9 @@ import path from "path";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
 
+const ReactCompilerConfig = {
+    target: "18"
+};
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -34,5 +37,12 @@ export default defineConfig(() => ({
             "@firecms/data_import_export": path.resolve(__dirname, "../data_import_export/src"),
         }
     },
-    plugins: [react({})]
+    plugins: [
+        react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })]
 }));