+++ packages/collection_editor_firebase/vite.config.ts
@@ -3,7 +3,9 @@ import path from "path";
 
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
-
+const ReactCompilerConfig = {
+    target: "18"
+};
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -40,5 +42,11 @@ export default defineConfig(() => ({
             "@firecms/data_import_export": path.resolve(__dirname, "../data_import_export/src"),
         }
     },
-    plugins: [react({})]
+    plugins: [react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })]
 }));