+++ packages/data_enhancement/vite.config.ts
@@ -3,7 +3,9 @@ import path from "path";
 
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
-
+const ReactCompilerConfig = {
+    target: "18"
+};
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -30,5 +32,11 @@ export default defineConfig(() => ({
             "@firecms/editor": path.resolve(__dirname, "../editor/src"),
         }
     },
-    plugins: [react()]
+    plugins: [react({
+        babel: {
+            plugins: [
+                ["babel-plugin-react-compiler", ReactCompilerConfig],
+            ],
+        }
+    })]
 }));