+++ packages/firebase_firecms/vite.config.ts
@@ -4,6 +4,10 @@ import path from "path";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
 
+const ReactCompilerConfig = {
+    target: "18"
+};
+
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -36,6 +40,12 @@ export default defineConfig(() => ({
         }
     },
     plugins: [
-        react({})
+        react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })
     ]
 }));