+++ packages/firecms_cloud/vite.config.ts
@@ -4,6 +4,10 @@ import path from "path";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
 
+const ReactCompilerConfig = {
+    target: "18"
+};
+
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -42,5 +46,11 @@ export default defineConfig(() => ({
             "@firecms/datatalk": path.resolve(__dirname, "../datatalk/src"),
         }
     },
-    plugins: [react({})]
+    plugins: [react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })]
 }));