+++ packages/firecms_core/vite.config.ts
@@ -4,6 +4,10 @@ import path from "path";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
 
+const ReactCompilerConfig = {
+    target: "18"
+};
+
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -27,10 +31,16 @@ export default defineConfig(() => ({
         alias: {
             "@firecms/ui": path.resolve(__dirname, "../ui/src"),
             "@firecms/editor": path.resolve(__dirname, "../editor/src"),
-            "@firecms/formex": path.resolve(__dirname, "../formex/src"),
+            "@firecms/formex": path.resolve(__dirname, "../formex/src")
         }
     },
     plugins: [
-        react({})
+        react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })
     ]
 }));