+++ packages/formex/vite.config.ts
@@ -4,6 +4,10 @@ import path from "path";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react"
 
+const ReactCompilerConfig = {
+    target: "18"
+};
+
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -25,5 +29,11 @@ export default defineConfig(() => ({
             external: isExternal
         }
     },
-    plugins: [react({})]
+    plugins: [react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        })]
 }));