+++ packages/ui/vite.config.ts
@@ -5,6 +5,10 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import { viteStaticCopy } from "vite-plugin-static-copy";
 
+const ReactCompilerConfig = {
+    target: "18"
+};
+
 const isExternal = (id: string) => !id.startsWith(".") && !path.isAbsolute(id);
 
 export default defineConfig(() => ({
@@ -25,7 +29,13 @@ export default defineConfig(() => ({
         }
     },
     plugins: [
-        react({}),
+        react({
+            babel: {
+                plugins: [
+                    ["babel-plugin-react-compiler", ReactCompilerConfig],
+                ],
+            }
+        }),
         viteStaticCopy({
             targets: [
                 {